Restored usage of safer (although not thread-safe) readdir()
readdir_r() does not behave well with filename lengths greater than
NAME_MAX: they are truncated to the size of d_name struct field,
and are not NUL-terminated. This is a problem on NTFS, which allows
filenames with 255 UTF-16 characters (510 Unix chars).

Also extended inotify buffer to prevent unnecessary watches
re-registration on events with such filename lengths.